fix: Added Alert link to Case Details
