+++ frontend/src/features/cases/components/TasksDetailsModal.tsx
@@ -3,7 +3,6 @@ import { XMarkIcon } from '@heroicons/react/24/outline';
 import type { CaseRow } from './casesTable.utils';
 import CollaboratePanel from './view/CollaboratePanel';
 import TaskEvidenceTab from './view/TaskEvidenceTab';
-import LinkedItemsTab from './view/LinkedItemsTab';
 import InvestigationNotesTab from './view/InvestigationNotesTab';
 import InvestigationSummaryTab from './view/InvestigationsSummaryTab';
 import TaskDetailsTab from './view/TaskDetailsTab';
@@ -16,7 +15,6 @@ type ViewTabKey =
   | 'details'
   | 'evidence'
   | 'visualizations'
-  | 'linked'
   | 'tasks'
   | 'notes'
   | 'customer'
@@ -204,7 +202,7 @@ const TaskDetailsModal: React.FC<TaskDetailsModalProps> = ({
   return (
     <div className="fixed inset-0 z-50 flex items-start justify-center bg-black/30 p-4">
       <div className="mt-6 w-full max-w-5xl rounded-lg bg-white shadow-lg max-h-[85vh] flex flex-col">
-        {}
+        { }
         <div className="flex items-center justify-between px-6 py-4">
           <div className="flex items-center gap-3">
             <h3 className="text-lg font-semibold text-gray-900">
@@ -220,13 +218,12 @@ const TaskDetailsModal: React.FC<TaskDetailsModalProps> = ({
           </button>
         </div>
 
-        {}
+        { }
         {!showCollaborate && (
           <div className="flex items-center gap-2 px-6 pt-3">
             {(
               [
                 { key: 'details', label: 'Task Details' },
-                { key: 'linked', label: 'Linked Items' },
                 { key: 'evidence', label: 'Evidence' },
                 ...(shouldShowVisualizations === true
                   ? ([
@@ -245,11 +242,10 @@ const TaskDetailsModal: React.FC<TaskDetailsModalProps> = ({
                 onClick={() => {
                   setTab(t.key);
                 }}
-                className={`-mb-px rounded-t-md px-3 py-2 text-sm font-medium ${
-                  tab === t.key
-                    ? 'border-b-2 border-indigo-600 text-indigo-700'
-                    : 'text-gray-600 hover:text-gray-800'
-                }`}
+                className={`-mb-px rounded-t-md px-3 py-2 text-sm font-medium ${tab === t.key
+                  ? 'border-b-2 border-indigo-600 text-indigo-700'
+                  : 'text-gray-600 hover:text-gray-800'
+                  }`}
               >
                 {t.label}
               </button>
@@ -257,7 +253,7 @@ const TaskDetailsModal: React.FC<TaskDetailsModalProps> = ({
           </div>
         )}
 
-        {}
+        { }
         {/* Content */}
         <div className="px-6 py-5 overflow-y-auto flex-1">
           {showCollaborate ? (
@@ -296,17 +292,6 @@ const TaskDetailsModal: React.FC<TaskDetailsModalProps> = ({
                   />
                 </div>
               )}
-              <div style={{ display: tab === 'linked' ? 'block' : 'none' }}>
-                {row?.id && (
-                  <LinkedItemsTab
-                    caseId={row.id}
-                    onNavigateToCase={() => {
-                      onClose();
-                      onSwitchToCaseDetails?.();
-                    }}
-                  />
-                )}
-              </div>
               <div style={{ display: tab === 'notes' ? 'block' : 'none' }}>
                 <InvestigationNotesTab
                   task={tasks.filter((t) => t.task_id === selectedTask?.id)[0]}
@@ -344,7 +329,7 @@ const TaskDetailsModal: React.FC<TaskDetailsModalProps> = ({
           )}
         </div>
 
-        {}
+        { }
         <div className="flex items-center justify-end gap-3 border-t border-gray-200 px-6 py-4">
           <button
             type="button"
